Fix layer renaming in treeview, make layer selection persist during serialization.
