+++ sdkproject/Assets/Mapbox/Unity/Editor/MapManagerEditor.cs
@@ -67,43 +67,11 @@
 		{
 			get
 			{
-				return EditorPrefs.GetBool("MapManagerEditor_showMapLayers");
+				return EditorPrefs.GetBool(objectId + "MapManagerEditor_showMapLayers");
 			}
 			set
 			{
-				EditorPrefs.SetBool("MapManagerEditor_showMapLayers", value);
-			}
-		}
-
-		/// <summary>
-		/// Gets or sets a value to show or hide Vector section <see cref="T:Mapbox.Editor.MapManagerEditor"/>.
-		/// </summary>
-		/// <value><c>true</c> if show vector; otherwise, <c>false</c>.</value>
-		bool ShowLocationPrefabs
-		{
-			get
-			{
-				return EditorPrefs.GetBool("MapManagerEditor_showLocationPrefabs");
-			}
-			set
-			{
-				EditorPrefs.SetBool("MapManagerEditor_showLocationPrefabs", value);
-			}
-		}
-
-		/// <summary>
-		/// Gets or sets a value to show or hide Vector section <see cref="T:Mapbox.Editor.MapManagerEditor"/>.
-		/// </summary>
-		/// <value><c>true</c> if show vector; otherwise, <c>false</c>.</value>
-		bool ShowFeatures
-		{
-			get
-			{
-				return EditorPrefs.GetBool(objectId + "MapManagerEditor_showVector");
-			}
-			set
-			{
-				EditorPrefs.SetBool(objectId + "MapManagerEditor_showVector", value);
+				EditorPrefs.SetBool(objectId + "MapManagerEditor_showMapLayers", value);
 			}
 		}
 
@@ -118,12 +86,6 @@
 				EditorPrefs.SetBool(objectId + "MapManagerEditor_showPosition", value);
 			}
 		}
-
-		private GUIContent _requiredMapIdGui = new GUIContent
-		{
-			text = "Required Map Id",
-			tooltip = "For location prefabs to spawn the \"streets-v7\" tileset needs to be a part of the Vector data source"
-		};
 
 		private GUIContent mapIdGui = new GUIContent
 		{
@@ -234,7 +196,6 @@
 
 		void DrawMapLayerOptions()
 		{
-			//EditorGUI.indentLevel++;
 			var vectorDataProperty = serializedObject.FindProperty("_vectorData");
 			var layerProperty = vectorDataProperty.FindPropertyRelative("_layerProperty");
 			var layerSourceProperty = layerProperty.FindPropertyRelative("sourceOptions");
@@ -310,123 +271,8 @@
 			EditorGUILayout.Space();
 			ShowSepartor();
 
-			ShowLocationPrefabs = EditorGUILayout.Foldout(ShowLocationPrefabs, "POINTS OF INTEREST");
-			if (ShowLocationPrefabs)
-			{
-				if (sourceTypeValue != VectorSourceType.None && layerString.Contains(streets_v7))
-				{
-					GUI.enabled = false;
-					EditorGUILayout.TextField(_requiredMapIdGui, streets_v7);
-					GUI.enabled = true;
-					ShowSection(vectorDataProperty, "_locationPrefabsLayerProperties");
-				}
-				else
-				{
-					EditorGUILayout.HelpBox("In order to place location prefabs please add \"mapbox.mapbox-streets-v7\" to the data source in the FEATURES section.", MessageType.Error);
-				}
-			}
-
-			ShowSepartor();
-			ShowFeatures = EditorGUILayout.Foldout(ShowFeatures, "FEATURES");
-			if (ShowFeatures)
-			{
-				ShowSection(serializedObject.FindProperty("_vectorData"), "_layerProperty");
-			}
-		}
-
-		void PresetLocationBased(SerializedProperty unifiedMap)
-		{
-			//Set
-			//placement = atLocationCenter,
-			//scaling = custom
-			//turn off vector layers.
-			var mapOptionsProp = unifiedMap.FindPropertyRelative("mapOptions");
-			var vectorLayerProps = unifiedMap.FindPropertyRelative("vectorLayerProperties");
-			var placementType = mapOptionsProp.FindPropertyRelative("placementOptions.placementType");
-			var scalingType = mapOptionsProp.FindPropertyRelative("scalingOptions.scalingType");
-			var unitType = mapOptionsProp.FindPropertyRelative("scalingOptions.unitType");
-			var extentType = mapOptionsProp.FindPropertyRelative("extentOptions.extentType");
-			var vectorSourceType = vectorLayerProps.FindPropertyRelative("sourceType");
-
-			placementType.enumValueIndex = (int)MapPlacementType.AtLocationCenter;
-
-			scalingType.enumValueIndex = (int)MapScalingType.Custom;
-			unitType.enumValueIndex = (int)MapUnitType.meters;
-
-			extentType.enumValueIndex = (int)MapExtentType.CameraBounds;
-
-			vectorSourceType.enumValueIndex = (int)VectorSourceType.None;
-
-		}
-
-		void PresetWorldSimulator(SerializedProperty unifiedMap)
-		{
-			//Set
-			//placement = atLocationCenter,
-			//scaling = custom
-			//turn on vector layers.
-			var mapOptionsProp = unifiedMap.FindPropertyRelative("mapOptions");
-			var vectorLayerProps = unifiedMap.FindPropertyRelative("vectorLayerProperties");
-			var placementType = mapOptionsProp.FindPropertyRelative("placementOptions.placementType");
-			var scalingType = mapOptionsProp.FindPropertyRelative("scalingOptions.scalingType");
-			var unitType = mapOptionsProp.FindPropertyRelative("scalingOptions.unitType");
-			var extentType = mapOptionsProp.FindPropertyRelative("extentOptions.extentType");
-			var vectorSourceType = vectorLayerProps.FindPropertyRelative("sourceType");
-
-			placementType.enumValueIndex = (int)MapPlacementType.AtLocationCenter;
-
-			scalingType.enumValueIndex = (int)MapScalingType.Custom;
-			unitType.enumValueIndex = (int)MapUnitType.meters;
-
-			extentType.enumValueIndex = (int)MapExtentType.CameraBounds;
-
-			vectorSourceType.enumValueIndex = (int)VectorSourceType.MapboxStreets;
-		}
-
-		void PresetARTableTop(SerializedProperty unifiedMap)
-		{
-			//Set
-			//placement = atLocationCenter,
-			//scaling = custom
-			//turn on vector layers.
-			var mapOptionsProp = unifiedMap.FindPropertyRelative("mapOptions");
-			var vectorLayerProps = unifiedMap.FindPropertyRelative("vectorLayerProperties");
-			var placementType = mapOptionsProp.FindPropertyRelative("placementOptions.placementType");
-			var scalingType = mapOptionsProp.FindPropertyRelative("scalingOptions.scalingType");
-			var extentType = mapOptionsProp.FindPropertyRelative("extentOptions.extentType");
-			var vectorSourceType = vectorLayerProps.FindPropertyRelative("sourceType");
-
-			placementType.enumValueIndex = (int)MapPlacementType.AtLocationCenter;
-
-			scalingType.enumValueIndex = (int)MapScalingType.WorldScale;
-
-			extentType.enumValueIndex = (int)MapExtentType.CameraBounds;
-
-			vectorSourceType.enumValueIndex = (int)VectorSourceType.MapboxStreets;
-
-		}
-
-		void PresetARWorldScale(SerializedProperty unifiedMap)
-		{
-			//Set
-			//placement = atLocationCenter,
-			//scaling = custom
-			//turn on vector layers.
-			var mapOptionsProp = unifiedMap.FindPropertyRelative("mapOptions");
-			var vectorLayerProps = unifiedMap.FindPropertyRelative("vectorLayerProperties");
-			var placementType = mapOptionsProp.FindPropertyRelative("placementOptions.placementType");
-			var scalingType = mapOptionsProp.FindPropertyRelative("scalingOptions.scalingType");
-			var extentType = mapOptionsProp.FindPropertyRelative("extentOptions.extentType");
-			var vectorSourceType = vectorLayerProps.FindPropertyRelative("sourceType");
-
-			placementType.enumValueIndex = (int)MapPlacementType.AtLocationCenter;
-
-			scalingType.enumValueIndex = (int)MapScalingType.WorldScale;
-
-			extentType.enumValueIndex = (int)MapExtentType.CameraBounds;
-
-			vectorSourceType.enumValueIndex = (int)VectorSourceType.MapboxStreets;
-
+			GUILayout.Space(-2.0f * _lineHeight);
+			ShowSection(serializedObject.FindProperty("_vectorData"), "_layerProperty");
 		}
 	}
 }
