--- sdkproject/Assets/Mapbox/Unity/Editor/PropertyDrawers/LocationPrefabsLayerPropertiesDrawer.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 8b1e7d6ffe30243af9d1fbe886f180b1
-timeCreated: 1523475002
-licenseType: Pro
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
