--- sdkproject/Assets/Mapbox/Unity/Editor/VectorSubLayerTreeView.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: a9591575401b14570a6a95b275639eaf
-timeCreated: 1518069137
-licenseType: Pro
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
