--- sdkproject/Assets/Mapbox/Unity/LayerProperties/LocationPrefabLayerProperties.cs
@@ -1,10 +1,0 @@
-using UnityEngine;
-using System.Collections.Generic;
-using Mapbox.Unity.Map;
-using System;
-
-[Serializable]
-public class LocationPrefabsLayerProperties : LayerProperties
-{
-	public List<PrefabItemOptions> locationPrefabList = new List<PrefabItemOptions>();
-}
