--- sdkproject/Assets/Mapbox/Unity/LayerProperties/LocationPrefabLayerProperties.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 3ee13521b77df4464a6e8571a5396a7d
-timeCreated: 1523396148
-licenseType: Pro
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
