+++ sdkproject/Assets/Mapbox/Unity/SourceLayers/VectorLayer.cs
@@ -10,10 +10,6 @@
 	public class VectorLayer : IVectorDataLayer
 	{
 		[SerializeField]
-		LocationPrefabsLayerProperties _locationPrefabsLayerProperties = new LocationPrefabsLayerProperties();	
-
-
-		[SerializeField]
 		VectorLayerProperties _layerProperty = new VectorLayerProperties();
 
 		[NodeEditorElement(" Vector Layer ")]
@@ -109,20 +105,20 @@
 		public void AddLocationPrefabItem(PrefabItemOptions prefabItem)
 		{
 			//ensure that there is a list of prefabitems
-			if (LocationPrefabsLayerProperties.locationPrefabList == null)
+			if (PointsOfInterestSublayerList == null)
 			{
-				LocationPrefabsLayerProperties.locationPrefabList = new List<PrefabItemOptions>();
+				PointsOfInterestSublayerList = new List<PrefabItemOptions>();
 			}
 
-			if(_layerProperty.locationPrefabList == null)
+			if (_layerProperty.locationPrefabList == null)
 			{
 				_layerProperty.locationPrefabList = new List<PrefabItemOptions>();
 			}
 
 			//add the prefab item if it doesn't already exist
-			if (!LocationPrefabsLayerProperties.locationPrefabList.Contains(prefabItem))
+			if (!PointsOfInterestSublayerList.Contains(prefabItem))
 			{
-				LocationPrefabsLayerProperties.locationPrefabList.Add(prefabItem);
+				PointsOfInterestSublayerList.Add(prefabItem);
 			}
 
 			//add the prefab item if it doesn't already exist
@@ -142,9 +138,9 @@
 
 		public void RemovePrefabItem(int index)
 		{
-			if (LocationPrefabsLayerProperties.locationPrefabList != null)
+			if (PointsOfInterestSublayerList != null)
 			{
-				LocationPrefabsLayerProperties.locationPrefabList.RemoveAt(index);
+				PointsOfInterestSublayerList.RemoveAt(index);
 			}
 		}
 
@@ -158,7 +154,7 @@
 		public void Initialize()
 		{
 			_vectorTileFactory = ScriptableObject.CreateInstance<VectorTileFactory>();
-			_layerProperty.locationPrefabList = LocationPrefabsLayerProperties.locationPrefabList;
+			_layerProperty.locationPrefabList = PointsOfInterestSublayerList;
 			_vectorTileFactory.SetOptions(_layerProperty);
 		}
 
@@ -184,11 +180,15 @@
 		}
 		private VectorTileFactory _vectorTileFactory;
 
-		public LocationPrefabsLayerProperties LocationPrefabsLayerProperties
+		public List<PrefabItemOptions> PointsOfInterestSublayerList
 		{
 			get
 			{
-				return _locationPrefabsLayerProperties;
+				return _layerProperty.locationPrefabList;
+			}
+			set
+			{
+				_layerProperty.locationPrefabList = value;
 			}
 		}
 	}
